# backend/docs/swagger.yaml
basePath: /api
definitions:
  controllers.AddToCartRequest:
    properties:
      product_id:
        example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
        type: string
      quantity:
        example: 2
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
  controllers.CategoryInput:
    properties:
      description:
        example: อาหารคุณภาพสูงสำหรับสุนัขทุกวัย
        type: string
      image_url:
        example: https://example.com/category.jpg
        type: string
      name:
        example: อาหารสุนัข
        type: string
    required:
    - name
    type: object
...
  controllers.UpdateCartRequest:
    properties:
      quantity:
        example: 3
        minimum: 1
        type: integer
    required:
    - quantity
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@petfood.com
    name: API Support
  description: This is a Pet Food E-commerce API server with authentication, products,
    cart, and order management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Pet Food E-commerce API
  version: "1.0"
paths:
  /admin/categories:
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/controllers.CategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new category (Admin only)
      tags:
      - Admin - Categories
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a category from the system
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a category (Admin only)
      tags:
      - Admin - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category's information
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/controllers.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a category (Admin only)
      tags:
      - Admin - Categories
...
  /cart:
    delete:
      consumes:
      - application/json
      description: Remove all items from the authenticated user's cart
      produces:
      - application/json
      responses:
        "200":
          description: Cart cleared successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Clear entire cart
      tags:
      - Cart
    get:
      consumes:
      - application/json
      description: Get all items in the authenticated user's shopping cart with total
      produces:
      - application/json
      responses:
        "200":
          description: Cart items and total
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user's cart
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: Add a product to the authenticated user's shopping cart
      parameters:
      - description: Cart item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cart updated (item already exists)
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Item added to cart
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or insufficient stock
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - Cart
  /cart/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a specific item from the cart
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Item removed from cart
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart item not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: Update the quantity of an item in the cart
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: string
      - description: New quantity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cart updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request or insufficient stock
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Cart item not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update cart item quantity
      tags:
      - Cart
  /categories:
    get:
      consumes:
      - application/json
      description: Get a list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: List of categories
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all categories
      tags:
      - Categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Category not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a category by ID
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"